subscribe + mfix example
I was wondering if we should pass `Subscription` as an extra parameter to
subscribe, to make it possible to unsubscribe from withing the handler, e.g.:

    CDP.pageEnable handle
    sub1 <- CDP.subscribe handle $ \sub e -> do
        putStrLn $ "1: " ++ CDP.pageFrameUrl (CDP.pageFrameNavigatedFrame e)
        CDP.unsubscribe handle sub

However, most of the time you want to ignore this parameter, so this makes it a
bit ugly.  But it turns out you can also do this using `mfix`, so I wanted to
add an example of that.